backend/requirements.txt
flask==2.3.3
flask-cors
flask-jwt-extended
mysql-connector-python
